Add parameter for obsolete sshd configuration options
Certain SSH options are obsolete in that even if you set the value,
OpenSSH is going to ignore it. This is true for UsePrivilegeSeparation,
Protocol, and RhostsRSAAuthentication options.

This commit updates the SSH oval templates to add another template
variable called `obsolete_since_version` which gives us the flexibility
of saying "only run this rule if the system has a version of OpenSSH
less than version X". Otherwise, the check will short-circuit since
OpenSSH is defaulting to a more secure configuration.

This makes the rules more accurate, even if a user is setting something
that doesn't make sense at run time (e.g., Protocol 1).